+++ docs/changelog.rst
@@ -13,6 +13,7 @@ CHANGELOG
 
 - Filter by multiple structures on Blades list (#3646)
 - Add a multiselect to filter the Blades by more than one manager
+- Filter by begin date by default on touristic events in APIv2 (#3597)
 
 2.99.0 (2023-07-18)
 -----------------------